Refactor templating functions into separate php partial
